Require puppet-lint 5.1+
This pulls in three bigger changes in puppet-lint:
* Avoid using the manifest lexer on YAML [#239](https://github.com/puppetlabs/puppet-lint/pull/239) (tokenrove)
* Add abstract data types to lexer type tokens list [#240](https://github.com/puppetlabs/puppet-lint/pull/240) (kenyon)
* Fix Ensure check methods can't modify tokens array and add Ruby 3.4 testing [#233](https://github.com/puppetlabs/puppet-lint/pull/233) (alexjfisher)

In a previous PR, https://github.com/voxpupuli/voxpupuli-puppet-lint-plugins/pull/64, we switched to Ruby 3.2 as a minimal Ruby version.
The new puppet-lint release did a Ruby version bump.
We don't want to end up in a situation where different users have
different ruby versions, which leads to different puppet-lint versions,
which leads to different linting of your puppet code.

Because of that, we cannot allow puppet-lint 4 & 5, but have to pint to
5.1 (5.0 doesn't support Ruby 3.4).

All puppet-lint plugins were also updated for puppet-lint 5.1, which
also required major releases.